Fix some problems and test driver

+++ Driver/Inc/driver_gpio.h
@@ -10,14 +10,17 @@
 
 #include "driver.h"
 
+/*
+ * This is a Configuration structure for a GPIO pin
+ */
 typedef struct
 {
-	uint8_t GPIO_PinNumber;
-	uint8_t GPIO_PinMode;
-	uint8_t GPIO_PinSpeed;
-	uint8_t GPIO_PinPuPControl;
-	uint8_t GPIO_PinOPType;
-	uint8_t GPIO_PinAltFunMode;
+	uint8_t GPIO_PinNumber;			/*!< 	possible values from @GPIO_PIN_NUMBERS 		>*/
+	uint8_t GPIO_PinMode;			/*!< 	possible values from @GPIO_PIN_MODES 		>*/
+	uint8_t GPIO_PinSpeed;			/*!< 	possible values from @GPIO_PIN_SPEED 		>*/
+	uint8_t GPIO_PinPuPControl;		/*!< 	possible values from @GPIO_PUPD_CONTROL 	>*/
+	uint8_t GPIO_PinOPType;			/*!< 	possible values from @GPIO_OP_TYPES 		>*/
+	uint8_t GPIO_PinAltFunMode;		/*!< 	possible values from @ 1 to 15 				>*/
 }GPIO_Pin_Config_t;
 
 /*********************Handle structure GPIO PIN*********************/
@@ -27,7 +30,10 @@ typedef struct
 	GPIO_Pin_Config_t 		GPIO_PinConfig;
 }GPIO_Handle_t;
 
-/**************************GPIO pin NUMBER***************************/
+/*
+ * @GPIO_PIN_NUMBERS
+ * GPIO pin numbers
+ */
 #define GPIO_PIN_0 			0
 #define GPIO_PIN_1 			1
 #define GPIO_PIN_2 			2
@@ -45,7 +51,10 @@ typedef struct
 #define GPIO_PIN_14			14
 #define GPIO_PIN_15			15
 
-/**************************GPIO pin Mode***************************/
+/*
+ * @GPIO_PIN_MODES
+ * GPIO pin possible modes
+ */
 #define GPIO_MODE_INPUT 	0
 #define GPIO_MODE_OUTPUT 	1
 #define GPIO_MODE_ALTFUNC	2
@@ -54,17 +63,26 @@ typedef struct
 #define GPIO_MODE_IT_RT		5
 #define GPIO_MODE_IT_RFT	6
 
-/**********************GPIO pin OUTPUT Type***********************/
+/*
+ * @GPIO_OP_TYPES
+ * GPIO pin possible output types
+ */
 #define GPIO_OP_TYPE_PP 	0
 #define GPIO_OP_TYPE_OD 	1
 
-/**********************GPIO pin OUTPUT Speed***********************/
+/*
+ * @GPIO_PIN_SPEED
+ * GPIO pin possible output speeds
+ */
 #define GPIO_SPEED_LOW		0
 #define GPIO_SPEED_MEDIUM	1
 #define GPIO_SPEED_FAST		2
 #define GPIO_SPEED_HIGH		3
 
-/**********************GPIO pin PUP AND PDOW***********************/
+/*
+ * @GPIO_PUPD_CONTROL
+ * GPIO pin pull up AND pull down configuration macros
+ */
 #define GPIO_NO_PUPD		0
 #define GPIO_PIN_PU			1
 #define GPIO_PIN_PD			2
@@ -88,7 +106,7 @@ void GPIO_ToggleOutputPin(GPIO_Reg_t *pGPIOx, uint8_t PinNumber);
 
 /*******************************ISR*******************************/
 void GPIO_IRQITConfig(uint8_t IRQNumber, uint8_t En_or_Di);
-void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint8_t IRQPriority);
+void GPIO_IRQPriorityConfig(uint8_t IRQNumber, uint32_t IRQPriority);
 
 void GPIO_IRQHandling(uint8_t PinNumber);
